Represent heat value with floats.
- This allows for more granularity in heat production, which is neccesary for easily balancing heat production

+++ aestuscraft_common/dk/kiljacken/aestuscraft/api/heat/IHeatConsumer.java
@@ -24,6 +24,6 @@ public interface IHeatConsumer {
      *            The amount to supply
      * @return The actual amount consumed by the heat consumer
      */
-    public int supplyHeat(int amount);
+    public float supplyHeat(float amount);
 
 }
